Changed dt to float

+++ Camera3D.cpp
@@ -14,7 +14,7 @@ void init_camera(Camera3D* cam, vec3 cam_pos, vec3 target_pos){
 	cam->turn_speed = 100;
 }
 
-void update_camera(Camera3D* cam, CameraMode cam_mode, GameInput &game_input, vec3 player_pos, double dt)
+void update_camera(Camera3D* cam, CameraMode cam_mode, GameInput &game_input, vec3 player_pos, float dt)
 {
     if(cam_mode == CAM_MODE_DEBUG)
     {
